requirements.txt
requests[security]==2.18.4
Events==0.3
websocket-client>=0.47.0
signalr-client-threads
